Streamlined GitlabCI pipeline
Rather than have separate ReImaging jobs specifically for the scheduled
triggered pipelines, they've been integrated into the 'push' jobs, where
the schedule can define the DOCKER_NO_CACHE variable to '--no-cache' to
force a rebuild of the image.

Push jobs also no longer make images postfixed with '-master' as these
images are about to be pushed to the registry anyways. Also makes their
use a bit easier locally. Thus, only non-master builds will have branch
name as a postfix.

Also removed are the except/only tags for schedules, as now it will be
only based on whether it's the master branch or not.

Lastly, on master builds, the projects that depend on this one will now
be triggered for building upon successful master builds here, currently
just the 'build-stec' project.