Added openSUSE images
In addition, reworked the image builder script to derive the image tags
and source image names from the directory rather than explicitly given
as an environment variable.

This does mean that more images will be built, but they will all be
built on a per-directory basis, so re-use will be maximally applied when
possible.

Removed the two separate separators for image/tag names, simplified so
that the customization and information is entirely within the tag
itself, simplifying the process.