Majorly refactored images
Merged all image dockerfiles to one. All compilers and tools are in all
image variants, and the choosing of the compiler and Conan setup is
performed on container start through the image's entrypoint script.

This ultimately means that only one large image needs to be downloaded
instead of the 2/3 before, as well as having a single image to extend
with new libraries.