Fixed entrypoint returning false positives
Bad versions of the images have been uploaded, after passing the CI/CD
pipeline.

The entrypoint script has always returned a success regardless of any
issues, which has accidentally allowed bad versions of the images to be
uploaded, as the pipeline checked for a successful return of a basic run
of that script.

To fix, I've set 'set -e' to the script, so that upon an error, which
was typically due to mismatching versions of python packages, it exits
with an error code.

Thus, the pipeline should now fail on bad image creation, and not upload
it.

As a small aside, changed the shebang for the entrypoint to use
'/usr/bin/env sh'.