Changed how platforms are determined for Linux
Previously builds were dictated at the CI level, with the specified
options being limited to the minimal set that worked for the whole set
of OS images.

Now it has been changed to use a per-Dockerfile confiog file of the same
base name. This means all platforms can be built for all the distros.

For example riscv is only available on some of the latest versions of
Ubuntu, but not earlier.