Removed ppc64 from all images
Unfortunately, on the newer docker images for these distros perform
illegal instructions on ppc64le causing image builds to fail.

This applies both to newer and older version of qemu from the 2018-2021
range. It is most probably an issue with docker choosing an older POWER
arch to emulate that doesn't contain these instructions.