Add new Windows versions
Now that there is more than one variant of Windows, the different tags
have taken on the same naming scheme that Linux has had, being
windows-<version> and a 'windows' tag being the same as the latest
variant, currently windows-2022.

The new windows-2022 uses the Visual Studio released in the same year.

Updated choco installs to not spam the log with download progress.