Rework CI pipelines for better flexibility
The current pipelines were useful, however suffered limitations where
all architectural variants of any image had to be created in the same
job, and eventually pushed with the understanding that the image was
still in the cache of the same runner.

This was acceptable since on Linux, other architectures could be done
via QEMU, and Windows images were amd64 only.

However there is a desire to start generating arm64 Windows images, and
other Linux machines running on different architectures are being
acquired that can build certain variants faster than by emulation.

Thus, the pipelines have been reworked with the use of an intermediate
private image registry, so that the first stage can build the variants
of each image, upload them all to the same location, then a second stage
can run that pulls the images together into the final images and pushed
to the final destination.

As well, there were issues when the same image was tagged differently on
Windows, where when running a 'no-cache' build, each re-tag was a
completely different image, instead of the same newly-built image being
re-tagged. This also fixes that by only re-tagging at the final push
stage, if ALTERNATE_TAGS is specified in the dockerfile.